+++ CHANGELOG.md
@@ -71,6 +71,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `Range:size/1`
 - Added missing `ledc` functions for esp32 platform
 - Added `network:connect/0,1` and `network:disconnect/0` to ESP32 network driver.
+- Added `network:sta_status/0` to get the current connection state of the sta interface.
 
 ### Changed